Moved database collections to seperate file
