+++ commands/general/rolepersist.js
@@ -1,5 +1,5 @@
 const { Command } = require('discord.js-commando');
-const { userData } = require('../../index');
+const { userData } = require('../../util/dbCollections');
 const { stripIndents } = require('common-tags');
 const Discord = require('discord.js');
 
@@ -16,6 +16,7 @@ module.exports = class RolePersistCommand extends Command {
                     key: 'command',
                     prompt: 'What role persist command do you want to perform?',
                     type: 'string',
+                    default: '',
                 }, 
                 {
                     key: 'member',
@@ -34,27 +35,38 @@ module.exports = class RolePersistCommand extends Command {
     }
 
     run(message, { command, member, source }) {
-        if (command === 'update') {
+        switch(command) {
+          case 'update':
             message.guild.members.cache.forEach(m => {
-                userData.updateRoles(m.user.id, message.guild.id, m.roles.cache.map(r => r.id));
+                userData.updateRoles(m.user.id, message.guild.id, 
+                                        m.roles.cache.map(r => r.id));
             });
-
             return message.say('Updated persistent roles database!');
-        } else if (command === 'list' || command === 'ls') {
+
+          case 'ls':
+          case 'list': {
             const taggedMember = member || message.member;
             let memberRoles = '';
 
-            const roles = userData.getRoles(taggedMember.user.id, message.guild.id);
+            const roles = 
+                userData.getRoles(taggedMember.user.id, message.guild.id);
+
             if (roles !== 'no roles') {
                 roles.forEach(r => {
                     if (message.guild.roles.cache.get(r)) {
-                        memberRoles += `${message.guild.roles.cache.get(r).name}\n`;
+                        memberRoles += 
+                            `${message.guild.roles.cache.get(r).name}\n`;
                     }
                 });
             }
 
-            return message.say(`${taggedMember.user.tag} roles:\n${Discord.Util.removeMentions(memberRoles)}`);
-        } else if (command === 'set') {
+            return message.say(stripIndents`
+                ${taggedMember.user.tag} roles:
+                ${Discord.Util.removeMentions(memberRoles)}
+            `);
+          }
+
+          case 'set': {
             let roleMember, updateMember;
             if (source) {
                 roleMember = source;
@@ -63,21 +75,25 @@ module.exports = class RolePersistCommand extends Command {
                 roleMember = member || message.member;
             }
            
-            const roles = userData.getRoles(roleMember.user.id, message.guild.id);
+            const roles = 
+                userData.getRoles(roleMember.user.id, message.guild.id);
             if (roles !== 'no roles') updateMember.roles.set(roles);
 
             return message.say(`Set roles for ${updateMember.user.tag}`);
-        } else if (command === 'help') {
+          }
+
+          case 'download': {
+            return message.channel.send({ files: ['./database.sqlite'] });
+          }
+
+          case 'help':
+          default:
             return message.say(stripIndents`
-                Help: 
+                **Role Persistence Commands**
                 \`update\` - updates role persist database for all members in a guild
                 \`set <destination_member> <source_member> \` - sets roles of destination with source
-                \`list <member>\` - lists the roles stored for role persistence of a member.
-            `); 
-        } else if (command === 'download') {
-            return message.channel.send({ files: ['./database.sqlite'] });
-        } else {
-            return message.say('Available commands: \`update\` \`set\` \`list\`');
+                \`list <member>\` - lists the roles stored for role persistence of a member
+            `);
         }
     }
 };
