+++ events/guildMemberAdd.js
@@ -1,4 +1,4 @@
-const { userData } = require('../index');
+const { userData } = require('../util/dbCollections');
 
 module.exports = {
 	name: 'guildMemberAdd',
