+++ index.js
@@ -1,13 +1,8 @@
 // discord.js-command module
 const { CommandoClient } = require('discord.js-commando');
-const { Collection } = require('discord.js');
 const { prefix, token, ownerID } = require('./config.json');
 const path = require('path');
 const fs = require('fs');
-
-const { Users } = require('./util/dbObjects');
-const userData = new Collection();
-module.exports = { userData };
 
 // console timestamps
 require('console-stamp')(console, { pattern: 'm/dd/yy HH:MM:ss', label: false, colors: { stamp: 'green' } });
@@ -40,62 +35,6 @@ client.dispatcher.addInhibitor(msg => {
     if(msg.member.roles.cache.some(r => r.name === 'bonk-mute')) return 'bonked';
 });
   
-Reflect.defineProperty(userData, 'updateRoles', {
-    value: async function updateRoles(user_id, guild_id, role_ids) {
-            const id = `${guild_id}-${user_id}`;    
-            const user = userData.get(id);
-            if (user) {
-                user.roles = role_ids;
-                return user.save();
-            }
-            const newUser = await Users.create({ 
-                id: id,
-                user_id: user_id, 
-                guild_id: guild_id,
-                roles: role_ids,
-            });
-            userData.set(id, newUser); 
-            return newUser;
-        },
-});
-
-Reflect.defineProperty(userData, 'getRoles', {
-    value: function getRoles(user_id, guild_id) {
-        const id = `${guild_id}-${user_id}`;
-        const user = userData.get(id);
-        return user ? user.roles : 'no roles';
-    },
-});
-
-Reflect.defineProperty(userData, 'addBalance', {
-	/* eslint-disable-next-line func-name-matching */
-	value: async function addBalance(user_id, guild_id, amount) {
-        const id = `${guild_id}-${user_id}`;
-        const user = userData.get(id);
-		if (user) {
-			user.balance += Number(amount);
-			return user.save();
-		}
-		const newUser = await Users.create({ 
-            id: id,
-            user_id: user_id, 
-            guild_id: guild_id,
-            balance: amount,
-        });
-		userData.set(id, newUser);
-		return newUser;
-	},
-});
-
-Reflect.defineProperty(userData, 'getBalance', {
-	/* eslint-disable-next-line func-name-matching */
-	value: function getBalance(user_id, guild_id) {
-        const id = `${guild_id}-${user_id}`;
-		const user = userData.get(id);
-		return user ? user.balance : 0;
-	},
-});
-
 // event handler
 const eventFiles = fs.readdirSync('./events').filter(file => file.endsWith('.js'));
